Catapult: Refactor Card
