+++ app/src/main/java/org/lineageos/tv/launcher/view/AddFavoriteCard.kt
@@ -13,30 +13,18 @@ import androidx.appcompat.content.res.AppCompatResources
 import org.lineageos.tv.launcher.R
 import org.lineageos.tv.launcher.model.Launchable
 
-class AddFavoriteCard : Card {
+class AddFavoriteCard @JvmOverloads constructor(
+    context: Context, attrs: AttributeSet? = null, defStyleAttr: Int = 0
+) : Card(context, attrs, defStyleAttr) {
     // Views
     private val actionIconView by lazy { findViewById<ImageView>(R.id.action_image) }
     private val iconView by lazy { findViewById<ImageView>(R.id.app_icon) }
     private val nameView by lazy { findViewById<TextView>(R.id.app_name) }
 
-    constructor(context: Context?) : super(context)
-
-    constructor(context: Context?, attrs: AttributeSet?) : super(context, attrs)
-
-    constructor(context: Context?, attrs: AttributeSet?, defStyleAttr: Int) : super(
-        context,
-        attrs,
-        defStyleAttr
-    )
-
     init {
         inflate(context, R.layout.favorites_add_app_card, this)
         background =
             AppCompatResources.getDrawable(context, R.drawable.favorites_app_card_background)
-    }
-
-    override fun inflate() {
-        inflate(context, R.layout.favorites_add_app_card, this)
     }
 
     fun setActionToggle(favorite: Boolean) {
